Script to make ConnectBox
This script is adapted from the prepare_release.py script and is essentially a sub-set of that script, doing the first task of checking for SSH connection and running the ansible script to put the required code on the target uSD card. It does not reserve a Release name on GitHub and ends when the loading of code to the target uSD is complete. This tool is well suited for experimenting and making local images for the ConnectBox. If desired, the shrinking and compressing of the image can then be manually completed using the shrink_image.py and xz tools. Should the user desire to release the image, this can be accomplished by manual interaction with GitHub.